requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
WTForms==3.1.2
SQLAlchemy==2.0.35
pymysql==1.1.1
python-dotenv==1.0.1
gunicorn==22.0.0
openpyxl==3.1.5
